mea_calibration_tool/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
numpy>=1.24.0
pillow>=10.0.0
rasterio>=1.3.0
